Gerando as tabelas automaticamente com JPA

+++ src/main/java/com/github/marcelofcamillo/libraryapi/model/Livro.java
@@ -3,12 +3,13 @@ package com.github.marcelofcamillo.libraryapi.model;
 import jakarta.persistence.*;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 
 @Entity
 @Table(name = "livro", schema = "public")
-@Data
+@Data // gera os getters e setters automáticamente
 public class Livro {
     @Id
     @Column(name = "id")
@@ -29,10 +30,10 @@ public class Livro {
     private GeneroLivro genero;
 
     @Column(name = "preco", precision = 18, scale = 2)
-    private Double preco;
+    private BigDecimal preco;
     //private BigDecimal preco;
 
-    @ManyToOne // muitos liberos para um autor
+    @ManyToOne // muitos livros para um autor
     @JoinColumn(name = "id_autor")
     private Autor autor;
 }
